Created UserConfirmation popup  and implement it on logout button

+++ src/components/LogoutButton.js
@@ -1,19 +1,25 @@
 import React from "react";
 import LogoutIcon from "../imgs/logout.png";
 import { useHistory } from "react-router-dom";
+import UserConfirmation from './UserConfirmation'
 
 const TEXT = "Logga ut";
 const LogoutButton = () => {
   const history = useHistory()
+
+  const handelLogout = () => {
+    UserConfirmation({
+      text:'Är du säker på att logga ut?',
+      confirmAction: () => {
+        localStorage.clear();
+        history.push('/');
+      }
+    })
+  }
   return (
     <button
       className="logoutButton"
-      onClick={() => {
-        /** show prombt */
-        localStorage.clear();
-        history.push('/');
-      }}
-
+      onClick={handelLogout}
     >
       <img src={LogoutIcon} alt="Exist"></img>
       {TEXT}
